fix(tests): faulty test case `sources_without_std` (#143)
One of the tests for the sysand sources command didn't actually check if standard libraries where filtered out properly, and as such missed the fact that code for adding old standard library hadn't been removed when support was dropped.

Remove the old code and make the test stricter to catch similar mishaps in the future.

Signed-off-by: victor.linroth.sensmetry <victor.linroth@sensmetry.com>